feat: Add gui implementation
- Add OpenGL and GLFW
- Make separate devel and shared libraries
