+++ src/main.cc
@@ -8,5 +8,15 @@ int main([[maybe_unused]] int argc, [[maybe_unused]] char *argv[])
 {
     fmt::print("Example {}\n", "project");
 
+    try {
+        yui::gui gui{"Yui", { 1280, 720 }};
+
+        gui.run([&gui] {
+            gui.main_menu().menu("Foo");
+        });
+    } catch (const std::exception& ex) {
+        fmt::print("Exception occured: {}", ex.what());
+    }
+
     return EXIT_SUCCESS;
 }
